+++ routes/index.js
@@ -3,7 +3,8 @@ var debug = require('debug')('kcapp:index');
 var express = require('express');
 var router = express.Router();
 
-var indexTemplate = require('../src/pages/index/index-template.marko');
+const template = require('marko');
+var indexTemplate = template.load(require.resolve('../src/pages/index/index-template.marko'));
 
 var axios = require('axios');
 var _ = require('underscore');
@@ -30,7 +31,7 @@ router.get('/', function (req, res, next) {
             offices: offices.data
         });
     })).catch(error => {
-        debug('Error when getting data for / ' + error);
+        debug(`Error when getting data for / ${error}`);
         next(error);
     });
 });