+++ CHANGELOG.md
@@ -7,14 +7,18 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 - Added new game type `Kill Bull`
 - Added support for stream of board for each player using external stream URLs
 - `Elo Changelog` page for player
+- Custom announcer voice for scores `100` to `180`
 
 #### Changed
 - Added `Outshot Type` and `Tic-Tac-Toe` board to Spectate page
 - Added `Bull` to `420` game type
 - Correctly handle leg finish when multiple people are connected
+- Updated to Marko 5
+- Updated all dependencies to latest version
 
 #### Fixed
 - Score input for `X01 Handicap`
+- Announce correct leg finished
 - Issue causing new-leg announcements to be played twice
 - Misc fixes and improvements